split out all routes
