+++ index.js
@@ -1,77 +1,19 @@
 const express = require('express');
 const helmet = require('helmet');
-const bcrypt = require('bcrypt');
-const db = require('./data/dbConfig');
-const Users = require('./users/users-model.js');
+
+const userRoutes = require('./users/user-routes');
+const registerRoutes = require('./register/register-routes');
 
 const server = express();
 
 server.use(helmet());
 server.use(express.json());
+server.use('/api/users', userRoutes);
+server.use('/api/register', registerRoutes);
 
 server.get('/', (req, res) => {
   res.status(200).json('Home route working');
 });
-
-server.post('/api/login', (req, res) => {
-  let { username, password } = req.body;
-
-  Users.findBy({ username })
-    .first()
-    .then(user => {
-      // check tha password guess against the database
-      if (user && bcrypt.compareSync(password, user.password)) {
-        res.status(200).json({ message: `Welcome ${user.username}!` });
-      } else {
-        res.status(401).json({ message: 'Invalid Credentials' });
-      }
-    })
-    .catch(error => {
-      res.status(500).json(error);
-    });
-});
-
-server.post('/api/register', async (req, res) => {
-  try {
-    const userInfo = req.body;
-    const hash = bcrypt.hashSync(userInfo.password, 4);
-    userInfo.password = hash;
-
-    db('users').insert(userInfo);
-    res.status(201).json(hash);
-  } catch (error) {
-    res.status(500).json({ error });
-  }
-});
-
-server.get('/api/users', restricted, (req, res) => {
-  Users.find()
-    .then(users => {
-      res.json(users);
-    })
-    .catch(err => res.send(err));
-});
-function restricted(req, res, next) {
-  const { username, password } = req.headers;
-
-  if (username && password) {
-    Users.findBy({ username })
-      .first()
-      .then(user => {
-        // check tha password guess against the database
-        if (user && bcrypt.compareSync(password, user.password)) {
-          next();
-        } else {
-          res.status(401).json({ message: 'You shall not pass!!' });
-        }
-      })
-      .catch(error => {
-        res.status(500).json(error);
-      });
-  } else {
-    res.status(401).json({ message: 'Please provide credentials' });
-  }
-}
 
 const port = process.env.PORT || 5000;
 server.listen(port, () =>
