+++ users/user-routes.js
@@ -1,0 +1,34 @@
+const user = require('express').Router();
+const bcrypt = require('bcrypt');
+const Users = require('../users/users-model.js');
+
+user.get('/', restricted, (req, res) => {
+  Users.find()
+    .then(users => {
+      res.json(users);
+    })
+    .catch(err => res.send(err));
+});
+function restricted(req, res, next) {
+  const { username, password } = req.headers;
+
+  if (username && password) {
+    Users.findBy({ username })
+      .first()
+      .then(user => {
+        // check tha password guess against the database
+        if (user && bcrypt.compareSync(password, user.password)) {
+          next();
+        } else {
+          res.status(401).json({ message: 'You shall not pass!!' });
+        }
+      })
+      .catch(error => {
+        res.status(500).json(error);
+      });
+  } else {
+    res.status(401).json({ message: 'Please provide credentials' });
+  }
+}
+
+module.exports = user;
